Extract Pin to src/pin.rs
- Move Pin code to pin module
- Tiny change to Display impl (print branch where applicable)
